+++ src/components/DaemonConfig/RegisterDaemonConfigModal.vue
@@ -3,16 +3,17 @@
 		<NcModal :show="show" @close="closeModal">
 			<div class="register-daemon-config-body">
 				<h2>{{ t('app_api', 'Register Deploy Daemon') }}</h2>
-				<NcNoteCard type="warning">
-					{{ t('app_api', 'Supported daemon accepts-deploy-id:') }}
-					<b>
-						<a href="https://cloud-py-api.github.io/app_api/DeployConfigurations.html" target="_blank">docker-install</a>
-					</b>,
-					<b>
-						<a href="https://cloud-py-api.github.io/app_api/tech_details/Deployment.html#manual-install-for-development" target="_blank">manual-install</a>
-					</b>
-				</NcNoteCard>
-				<p>{{ t('app_api', 'These are the default settings for a regular DaemonConfig. You can change them as you wish; more info on that is in the docs.') }}</p>
+				<div class="templates">
+					<div class="external-label">
+						<label for="daemon-template">{{ t('app_api', 'Daemon configuration template') }}</label>
+						<NcSelect
+							id="daemon-template"
+							v-model="configurationTab"
+							:label-outside="true"
+							:options="configurationTemplateOptions"
+							:placeholder="t('app_api', 'Select daemon configuration template')" />
+					</div>
+				</div>
 				<form class="daemon-register-form">
 					<div class="external-label">
 						<label for="daemon-name">{{ t('app_api', 'Name') }}</label>
@@ -101,6 +102,15 @@
 									:aria-label="t('app_api', 'Hostname to reach ExApp (optional)')"
 									:helper-text="t('app_api', 'Hostname to reach ExApp (optional)')" />
 							</div>
+							<div v-if="['http', 'https'].includes(protocol)" class="external-label">
+								<label for="deploy-config-haproxy-password">{{ t('app_api', 'HaProxy password') }}</label>
+								<NcInputField
+									id="deploy-config-haproxy-password"
+									:value.sync="deployConfig.haproxy_password"
+									:placeholder="t('app_api', 'AppAPI Docker Socket Proxy authentication password')"
+									:aria-label="t('app_api', 'AppAPI Docker Socket Proxy authentication password')"
+									:helper-text="t('app_api', 'AppAPI Docker Socket Proxy authentication password')" />
+							</div>
 							<NcCheckboxRadioSwitch
 								id="deploy-config-gpus"
 								:checked.sync="deployConfig.gpu"
@@ -110,7 +120,7 @@
 								{{ t('app_api', 'Enable GPUs support') }}
 							</NcCheckboxRadioSwitch>
 							<p v-if="deployConfig.gpu" class="hint">
-								{{ t('app_api', 'All GPU devices will be requested to be enabled in ExApp containers') }}
+								{{ t('app_api', 'All GPU devices will be requested to be enabled in ExApp containers by Docker') }}
 							</p>
 						</div>
 					</template>
@@ -119,6 +129,7 @@
 							{{ t('app_api', 'Register') }}
 							<template #icon>
 								<NcLoadingIcon v-if="registeringDaemon" :size="20" />
+								<Check v-else :size="20" />
 							</template>
 						</NcButton>
 					</div>
@@ -135,21 +146,21 @@ import { generateUrl } from '@nextcloud/router'
 
 import NcModal from '@nextcloud/vue/dist/Components/NcModal.js'
 import NcInputField from '@nextcloud/vue/dist/Components/NcInputField.js'
-import NcNoteCard from '@nextcloud/vue/dist/Components/NcNoteCard.js'
 import NcCheckboxRadioSwitch from '@nextcloud/vue/dist/Components/NcCheckboxRadioSwitch.js'
 
 import NcSelect from '@nextcloud/vue/dist/Components/NcSelect.js'
 import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
 
 import NcLoadingIcon from '@nextcloud/vue/dist/Components/NcLoadingIcon.js'
+import Check from 'vue-material-design-icons/Check.vue'
 import UnfoldLessHorizontal from 'vue-material-design-icons/UnfoldLessHorizontal.vue'
 import UnfoldMoreHorizontal from 'vue-material-design-icons/UnfoldMoreHorizontal.vue'
+import { DAEMON_TEMPLATES } from '../../constants/daemonTemplates.js'
 
 export default {
 	name: 'RegisterDaemonConfigModal',
 	components: {
 		NcLoadingIcon,
-		NcNoteCard,
 		NcModal,
 		NcInputField,
 		UnfoldLessHorizontal,
@@ -157,6 +168,7 @@ export default {
 		NcCheckboxRadioSwitch,
 		NcSelect,
 		NcButton,
+		Check,
 	},
 	props: {
 		show: {
@@ -184,15 +196,14 @@ export default {
 			deployConfig: {
 				net: 'host',
 				host: 'localhost',
-				ssl_key: '',
-				ssl_key_password: '',
-				ssl_cert: '',
-				ssl_cert_password: '',
+				haproxy_password: null,
 				gpu: false,
 			},
 			defaultDaemon: false,
 			registeringDaemon: false,
 			registerInOneClickLoading: false,
+			configurationTab: 'custom',
+			configurationTemplateOptions: ['custom', ...DAEMON_TEMPLATES.map(template => template.name)],
 		}
 	},
 	computed: {
@@ -206,6 +217,9 @@ export default {
 		},
 	},
 	watch: {
+		configurationTab(newConfigurationTab) {
+			this.setupFormConfiguration(newConfigurationTab)
+		},
 		acceptsDeployId(newAcceptsDeployId) {
 			if (newAcceptsDeployId === 'manual-install') {
 				this.name = 'manual_install'
@@ -224,6 +238,9 @@ export default {
 		},
 	},
 	methods: {
+		DAEMON_TEMPLATES() {
+			return DAEMON_TEMPLATES
+		},
 		registerDaemon() {
 			this.registeringDaemon = true
 			axios.post(generateUrl('/apps/app_api/daemons'), {
@@ -237,10 +254,6 @@ export default {
 						net: this.deployConfig.net,
 						host: this.deployConfig.host,
 						nextcloud_url: this.nextcloud_url,
-						ssl_key: this.deployConfig.ssl_key,
-						ssl_key_password: this.deployConfig.ssl_key_password,
-						ssl_cert: this.deployConfig.ssl_cert,
-						ssl_cert_password: this.deployConfig.ssl_cert_password,
 						gpu: this.deployConfig.gpu,
 					},
 				},
@@ -261,6 +274,28 @@ export default {
 					console.debug(err)
 					showError(t('app_api', 'Failed to register DaemonConfig. Check the logs'))
 				})
+		},
+		setupFormConfiguration(templateName) {
+			if (templateName === 'custom') {
+				this.setFormDefaults()
+				return
+			}
+			const template = DAEMON_TEMPLATES.find(template => template.name === templateName)
+			if (!template) {
+				return
+			}
+			this.name = template.name
+			this.displayName = template.displayName
+			this.acceptsDeployId = template.acceptsDeployId
+			this.protocol = template.protocol
+			this.host = template.host
+			this.nextcloud_url = template.nextcloud_url ?? window.location.origin + generateUrl('').slice(0, -1)
+			this.deployConfigSettingsOpened = template.deployConfigSettingsOpened
+			this.deployConfig.net = template.deployConfig.net
+			this.deployConfig.host = template.deployConfig.host
+			this.deployConfig.haproxy_password = template.deployConfig.haproxy_password
+			this.deployConfig.gpu = template.deployConfig.gpu
+			this.defaultDaemon = template.defaultDaemon
 		},
 		onProtocolChange() {
 			// Prefill default value
@@ -287,10 +322,7 @@ export default {
 			this.deployConfig = {
 				net: 'host',
 				host: 'localhost',
-				ssl_key: '',
-				ssl_key_password: '',
-				ssl_cert: '',
-				ssl_cert_password: '',
+				haproxy_password: null,
 				gpu: false,
 			}
 			this.registeringDaemon = false
@@ -335,6 +367,14 @@ export default {
 
 	.hint {
 		color: var(--color-warning-text)
+	}
+
+	.templates {
+		display: flex;
+		margin: 0 auto;
+		width: fit-content;
+		border-bottom: 1px solid var(--color-border-dark);
+		padding-bottom: 20px;
 	}
 }
 </style>